Fix asset loading technique
Moved asset loading to index.js to be handled before game begins. Previously, it was hacked into the creation of any Bullet which is bad news. Now, you define each asset in the index.js and store them in the assets object which is then iterated over, checking the .onload of each before starting the game.