Add cannons
Player is the keeper of the cannon graphic, and it might get the bullet graphic later on too. Fixed a problem with the drawRotatedImage function that I didn't realize wasn't working exactly as needed. With those changes in place, the cannons have been added, and the bullets have been shifted inward by the height of the cannon.