Add bomb drop timing
Created the structure needed for timed drops. There's a base interval (called the bombDelay), and a variance that can be set in the Game. The variance should be something between zero and 1 and effectively adjusts the possible delay between each bomb drop +/- that variance. So... a 1000 millisecond base with .2 variance will drop bombs separated by anywhere from 800 to 1200 milliseconds.
However, the variance doesn't seem obivous to me when I play, even when cranked up quite a bit. I'm going to pursue clustering the variances. Meaning, some small number of bombs dropped in a row will have the same delay, and then it'll change again, and some similar number of bombs dropped with keep that new delay. This should make the larger swings of variance be more obvious and force some periods of both rest and anxiety... I hope.