Add ability to put animated sprites on screen
Added drawAnimatedImage and drawCharacters in order to add Mario, Luigi, Yoshi, Peach, and Toad to the game. I also changed the image loader to have several new properties, like frames, frameInterval, and animationCycle. Also made the start screen a little prettier.
Also, I started working on something else because I forgot to commit! I'm changing the health system to reduce by 2 when a character is killed and still reduced by one if the character is already gone when a bomb hits.