Add bad guy drawing and better scoring
Worked in the infrastructure for drawing bad guys in addition to the good guys and added in some special scoring for when you kill the bad guys or just get a shot across. A kill is worth 50, and getting a shot all the way across is worth 20. Bullet kills are only worth 1 point.